Add the action bar
